gh-default: Default delivery as github
