+++ k8s/fn-task-controller/templates/controller-configmap.yaml
@@ -10,10 +10,11 @@ data:
   KC_HOST: http://keycloak.{{ include "kc_ns" . }}.svc.cluster.local
   GIT_HOME: {{ template "fn-task-controller.gitpath" . }}
   MOUNT_PATH: {{ (.Values.storage.local).path }}
-  PULL-POLICY: IfNotPresent
+  PULL-POLICY: Always
   TAG: {{ .Values.controller.tag | default .Chart.AppVersion }}
   IMAGE: {{ template "fn-task-controller.image" . }}
   PUBLIC_URL: {{ .Values.global.host | default "federated-node" }}
+  DEFAULT_DELIVERY_REPO: {{ (.Values.delivery.github).repo }}
 {{- if .Values.storage.azure }}
   AZURE_STORAGE_ENABLED: "true"
   AZURE_SHARE_NAME: {{ .Values.storage.azure.shareName }}/controller
