+++ k8s/fn-task-controller/values.yaml
@@ -40,3 +40,7 @@ idp:
 global: {}
   # host:
   # orgAndRepo:
+
+delivery:
+  github: {}
+    # repo:
